Add feed_meta helper for Ruby templates
