+++ lib/bridgetown-feed.rb
@@ -2,11 +2,5 @@
 
 require "bridgetown"
 require "fileutils"
+require "bridgetown-feed/builder"
 require "bridgetown-feed/generator"
-
-module BridgetownFeed
-  autoload :MetaTag,          "bridgetown-feed/meta-tag"
-  autoload :PageWithoutAFile, "bridgetown-feed/page-without-a-file.rb"
-end
-
-Liquid::Template.register_tag "feed_meta", BridgetownFeed::MetaTag
