--- lib/bridgetown-feed/page-without-a-file.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module BridgetownFeed
-  class PageWithoutAFile < Bridgetown::Page
-    def read_yaml(*)
-      @data ||= {}
-    end
-  end
-end
