+++ lib/bridgetown-feed/version.rb
@@ -2,6 +2,6 @@
 
 module Bridgetown
   module Feed
-    VERSION = "1.1.3"
+    VERSION = "2.0.0"
   end
 end
