+++ spec/fixtures/src/_layouts/helper.erb
@@ -1,0 +1,11 @@
+---
+---
+<html>
+  <head>
+    <%= feed_meta %>
+  </head>
+  <body>
+THIS IS MY ERB LAYOUT
+{{ content }}
+  </body>
+</html>
