+++ spec/fixtures/src/witherb.md
@@ -1,0 +1,6 @@
+---
+layout: helper
+title: I'm a page
+---
+
+Page content here
